wait for receiving
